fix: 기능별로 커스텀 훅 분리 피드백 적용
